API: rename 'resource_registry' -> 'assets'
