+++ databroker/assets/handlers_base.py
