+++ databroker/assets/json/AD_HDF5_datum.json
