+++ databroker/assets/json/AD_SPE_datum.json
