+++ databroker/assets/json/AD_SPE_resource.json
