+++ databroker/assets/path_only_handlers.py
