+++ databroker/assets/readers/spe.py
